Pizza Hut Sales
Welcome to the Sales Analysis of Pizza Hut, a comprehensive data analysis project utilizing SQL. This presentation covers a detailed examination of Pizza Hut's sales data, from fundamental queries to advanced analytical techniques. Our objective is to explore and interpret the sales trends, customer behavior, and performance metrics that drive Pizza Hut's success. Whether you are new to SQL or an experienced analyst, this project is designed to demonstrate the power and versatility of SQL in extracting valuable insights from complex datasets. Join us as we delve into the world of data to uncover actionable intelligence and strategic opportunities for Pizza Hut.